perf(ui): defer SessionSwitcher hub until open, split 2.2MB eager chunk
Extract SwitcherContent (useSwitcherItems, session stores, 2.2MB hub) to
SessionSwitcherContent.tsx and load it via React.lazy + Suspense only when
the dropdown opens. Header and MiniChat lazily import the wrapper itself so
the hub is not in the initial bootstrap graph. Vite now emits wrapper 8.5kB
+ content 8.4kB async chunks; the previous 2.2MB SessionSwitcherDropdown
eager chunk is eliminated. Main entry stays 146kB. Suspense fallbacks render
the trigger button immediately.

Validation: type-check + lint pass, build vendor-katex still lazy, no new
circular deps.
